% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDescriptiveCode.R
\name{generateDescriptiveCode}
\alias{generateDescriptiveCode}
\title{Generating markdown document with descriptive statistics code}
\usage{
generateDescriptiveCode(df)
}
\arguments{
\item{df}{df}
}
\value{
file
}
\description{
Generating markdown document with descriptive statistics code
}
\examples{










}
