% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histopath.R
\docType{package}
\name{histopathR}
\alias{histopathR}
\alias{histopathR-package}
\title{histopathR: Collection Of R Codes Used in Histopathological Research.}
\description{
Welcome to the histopathR package.
}
\details{
I have been collecting R Codes in another repository: [My R Codes For Data Analysis](\url{https://github.com/sbalci/MyRCodesForDataAnalysis}).
As I have stated before:
 > **My aim** is to collect all the codes one needs, where one starts with an excel or spss file and then end with the most common analysis used in histopathology papers : [example table](\url{https://media.nature.com/original/nature-assets/modpathol/journal/v31/n1/extref/modpathol2017106x4.docx})
 > There are plenty of ways to do an analysis in R, which is great but also confusing for the newbies. I will collect the codes here so that I can refer later and then update them as I learn more.
 
 In this package I would like to combine this effort under a package.
 
 If you would like to help me to collect R codes for histopathological data analysis please contact me, [issue a problem](\url{https://github.com/sbalci/histopathR/issues}) or feel free to [make a pull request](\url{https://github.com/sbalci/histopathR/pulls}).
}
\section{histopathR functions}{

The histopathR functions ...
}

