% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{Prices of 50,000 round cut diamonds.}
\format{gives an overview of the dataset. For data frames, you should include 
definition list that describes each variable. It’s usually a good idea to describe
variables’ units here
A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
x
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
