% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnumeric.R
\name{fnumeric}
\alias{fnumeric}
\title{Make numeric data from factor}
\usage{
fnumeric(x)
}
\arguments{
\item{x}{factor}
}
\value{
numeric
}
\description{
Make numeric data from factor
}
\examples{
a <- as.factor(c("2","4","6"))
as.numeric(a)
fnumeric(a)
}
