% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDescriptiveCode.R
\name{generateDescriptiveCode}
\alias{generateDescriptiveCode}
\title{Generating markdown document with descriptive statistics code}
\usage{
generateDescriptiveCode(dataframe, fileLokalization)
}
\arguments{
\item{dataframe}{dataframe}

\item{fileLokalization}{character}
}
\value{
file
}
\description{
Generating markdown document with descriptive statistics code
}
\examples{
generateDescriptiveCode <- function(dataframe, fileLokalization) {

magicfor::magic_for()

for (i in 1:length(names(dataframe))) {
  y <- paste0("```{r}
feopgl \%>\% 
  janitor::tabyl(", names(dataframe)[i], ") \%>\%
  adorn_pct_formatting(rounding = 'half up', digits = 1) \%>\%
  knitr::kable()
```", "\\n", "\\n")
  magicfor::put(y)
}

writeLines(magicfor::magic_result_as_vector(), fileLokalization)
magicfor::magic_free()
}



















}
