% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDescriptiveCode.R
\name{generateDescriptiveCode}
\alias{generateDescriptiveCode}
\title{generating code for descriptive statistics ----}
\usage{
generateDescriptiveCode(dataframe, fileLokalization)
}
\arguments{
\item{dataframe}{}

\item{fileLokalization}{}
}
\description{
generating code for descriptive statistics ----
}
